+++ lib/Peast/Syntax/Parser.php
@@ -3536,8 +3536,6 @@ class Parser extends \Peast\Syntax\ParserAbstract
                 return $literal;
             } elseif ($literal = $this->parseNumericLiteral()) {
                 return $literal;
-            } elseif ($literal = $this->parseBigIntLiteral()) {
-                return $literal;
             }
         }
         return null;
@@ -3565,7 +3563,7 @@ class Parser extends \Peast\Syntax\ParserAbstract
     /**
      * Parses a numeric literal
      * 
-     * @return Node\NumericLiteral|null
+     * @return Node\NumericLiteral|Node\BigIntLiteral|null
      */
     protected function parseNumericLiteral()
     {
@@ -3577,19 +3575,7 @@ class Parser extends \Peast\Syntax\ParserAbstract
             $node = $this->createNode("NumericLiteral", $token);
             $node->setRaw($val);
             return $this->completeNode($node);
-        }
-        return null;
-    }
-    
-    /**
-     * Parses a BigInt literal
-     * 
-     * @return Node\BigInt|null
-     */
-    protected function parseBigIntLiteral()
-    {
-        $token = $this->scanner->getToken();
-        if ($token && $token->getType() === Token::TYPE_BIGINT_LITERAL) {
+        } elseif ($token && $token->getType() === Token::TYPE_BIGINT_LITERAL) {
             $val = $token->getValue();
             $this->scanner->consumeToken();
             $node = $this->createNode("BigIntLiteral", $token);