+++ lib/Peast/Syntax/ES2020/Features.php
@@ -45,4 +45,11 @@ class Features extends \Peast\Syntax\ES2019\Features
      * @var bool
      */
     public $importMeta = true;
+
+    /**
+     * Coalescing operator
+     *
+     * @var bool
+     */
+    public $coalescingOperator = true;
 }